Add bypassCache callback to allow for not returning cached content
The bypassCache callback allows for finer-grained control of when it is
ok for a request to return cached content. It allows for the cached
version of the request to be ignored if bypassCache returns true. This
may be useful when only certain browsers or status codes should be
allowed to receive cached content.